// src/UrlTracker.Backoffice.UI/api-client/types.gen.ts
export type RedirectCollectionResponse = {
  results: Array<RedirectResponse>;
  total: number;
};

export type RedirectRequest = {
  source: StrategyViewModel;
  target: StrategyViewModel;
  permanent: boolean;
  retainQuery: boolean;
  force: boolean;
  advanced: boolean;
};

export type RedirectResponse = {
  source: StrategyViewModel;
  target: StrategyViewModel;
  permanent: boolean;
  retainQuery: boolean;
  force: boolean;
  advanced: boolean;
  id: number;
  key: string;
  createDate: string;
  updateDate: string;
  readonly additionalData: {
    [key: string]: unknown;
  };
};

export type RedirectType = 'None' | 'Temporary' | 'Permanent' | 'All';

export type ReferrerResponse = {
  referrerOccurances: number;
  referrerUrl: string;
};

export type ScoreParametersResponse = {
  variableFactor: number;
  redactionFactor: number;
  timeFactor: number;
};

export type StrategyViewModel = {
  strategy: string;
  value: string;
};